+++ freer.cabal
@@ -48,8 +48,6 @@ library
                      , Control.Monad.Freer.Writer
                      , Data.FTCQueue
                      , Data.Open.Union
-                     , Data.Open.Union51
-
   build-depends:       base >=4.7 && <5
   hs-source-dirs:      src
   ghc-options:         -Wall
